requirements.txt
tqdm
numpy
pandas
biopython
requests
